add tests for custom headers
